TodosApp/pyproject.toml
[tool.poetry]
name = "todos"
version = "0.1.0"
description = ""
authors = ["hanankhalid1 <hanankhalid1434@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.21"
uvicorn = "^0.30.6"
standard = "^1.0.3"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
fastapi = "^0.112.1"

[tool.poetry.scripts]
dev = "todos.main:start"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
